docs: rename OpenAPI to JSON Schema
