+++ .mergify.yml
@@ -63,7 +63,7 @@ queue_rules:
       - or:
           - and:
               - author=mergify-ci-bot
-              - head~=^openapi-spec-sync$
+              - head~=^configuration-spec-sync$
               - "title~=^chore: sync"
           - and:
               - author=mergify-ci-bot
