+++ src/components/Tables/ActionOptionsTable.tsx
@@ -1,4 +1,4 @@
-import configSchema from '../../../public/mergify-configuration-openapi.json';
+import configSchema from '../../../public/mergify-configuration-schema.json';
 
 import { OptionDefinition } from './ConfigOptions';
 import { OptionsTableBase } from './OptionsTable';
