+++ src/content/docs/configuration/file-format.mdx
@@ -155,16 +155,16 @@ Configuration Files](/configuration/sharing#extending-configuration-files).
 
 {/* <OptionsTable name="Extend" /> */}
 
-## OpenAPI Specification
+## JSON Schema Specification
 
 For those interested in a more detailed and machine-readable description of the
-Mergify configuration file format, an [OpenAPI](https://www.openapis.org/)
+Mergify configuration file format, an [JSON Schema](https://json-schema.org/)
 specification is available. This specification can help you understand the
 structure of the configuration file and can be used to generate client
 libraries, server stubs, or API documentation.
 
-You can find <DownloadLink href="/mergify-configuration-openapi.json">the
-OpenAPI specification for the Mergify configuration here</DownloadLink>.
+You can find <DownloadLink href="/mergify-configuration-schema.json">the JSON
+Schema specification for the Mergify configuration here</DownloadLink>.
 
 ## Validation and Troubleshooting
